Switch int to long in sales

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
 
     public static void main(String [] args) {
-        int [] sales={34200, 45000, 28600, 35800, 44800};
+        long [] sales={34200, 45000, 28600, 35800, 44800};
         SalesManager manager = new SalesManager(sales);
         System.out.println(manager.max());
     }
